Renamed command_definition.py to command_set.py.
