--- docs/api/command_definition.md
@@ -1,3 +1,0 @@
-# cmd2.command_definition
-
-::: cmd2.command_definition
