+++ docs/api/command_set.md
@@ -1,0 +1,3 @@
+# cmd2.command_set
+
+::: cmd2.command_set
